refactor: moved MissingConfigurationException to util package
