+++ teste/src/test/java/pages/impl/util/MissingConfigurationException.java
@@ -1,4 +1,4 @@
-package pages.impl;
+package pages.impl.util;
 
 public class MissingConfigurationException extends RuntimeException {
     public MissingConfigurationException(String message) {
