+++ teste/src/test/java/pages/impl/util/Utilities.java
@@ -1,8 +1,5 @@
 package pages.impl.util;
 
-import pages.impl.MissingConfigurationException;
-
-import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.Properties;
